requirements.txt
cachetools==4.1.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
Flask==1.1.2
get==2019.4.13
google-api-core==1.17.0
google-api-python-client==1.8.0
google-auth==1.14.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.51.0
grpcio==1.27.2
grpcio-tools==1.27.2
gspread==3.4.2
httplib2==0.17.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
oauth2client==4.1.3
oauthlib==3.1.0
post==2019.4.13
protobuf==3.11.3
public==2019.4.13
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydub==0.23.1
pytz==2019.3
query-string==2019.4.13
request==2019.4.13
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
selenium==3.141.0
six==1.14.0
soupsieve==2.0
uritemplate==3.0.1
urllib3==1.25.8
Werkzeug==1.0.1
youtube-dl==2020.3.8
youtube-search==0.1.3
